update: restructured onboarding parent to pass session and updated next config to include google and github images
