+++ next.config.mjs
@@ -1,4 +1,19 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  images: {
+    remotePatterns: [
+      {
+        protocol: 'https',
+        hostname: '**.githubusercontent.com/**',
+        port: '',
+      },
+      {
+        protocol: 'https',
+        hostname: '**.googleusercontent.com/**',
+        port: '',
+      },
+    ],
+  },
+};
 
 export default nextConfig;
